update fix bug

+++ src/pages/Home.jsx
@@ -9,9 +9,26 @@ export default function Home() {
   const [cartItems, setCartItems] = useState([]);
 
   const addToCart = (product) => {
-    const updatedCartItems = [...cartItems, product];
-    setCartItems(updatedCartItems);
-    localStorage.setItem("cartItems", JSON.stringify(updatedCartItems));
+    const existingItem = cartItems.find((item) => item.id === product.id);
+
+    if (existingItem) {
+      // Jika item sudah ada, perbarui jumlah produknya
+      const updatedCartItems = cartItems.map((item) =>
+        item.id === existingItem.id
+          ? { ...item, jumlah: item.jumlah + 1 }
+          : item
+      );
+
+      setCartItems(updatedCartItems);
+      localStorage.setItem("cartItems", JSON.stringify(updatedCartItems));
+      toast.success("Jumlah Berhasil Di Perbarui");
+    } else {
+      // Jika item belum ada, tambahkan item baru ke keranjang
+      const updatedCartItems = [...cartItems, { ...product, jumlah: 1 }];
+
+      setCartItems(updatedCartItems);
+      localStorage.setItem("cartItems", JSON.stringify(updatedCartItems));
+    }
   };
 
   const removeFromCart = (item) => {
